move file preview into own file
